Iss206 Removed use of gpg-agent from signing
This commit removed the use of gpg-agent, allowing the server optimised
gpg1 suite to be used rather then workstation designed gpg2.

Now, when run non-interativley the backup tool must be run with
passphrase-less signing keys.

When run interactively using a passhprase is still possible at the
console.

Also change the order of operations whereby the bucket directory was
being emptied prior to the gpg verification which created problems if
verification failed.